# IMap.putAll() performance improvements for ReplicatedMap