* Reduced litter in `ReplicatedMapEntries`

This is a partly adoption of https://github.com/hazelcast/hazelcast/pull/8023 for the `ReplicatedMap`.

I made a very short benchmark with 100 entries per user map for this PR:
```
50000 operations done in   44142 ms (+844 ms) (1184.83 ops/s) (844 µs per op)
(1132.71 ops/s avg) (882 µs per op avg)...

vs.

50000 operations done in   39366 ms (+807 ms) (1239.16 ops/s) (807 µs per op)
(1270.13 ops/s avg) (787 µs per op avg)...
```