@devOpsHazelcast 2016-05-18
Test PASSed.