@devOpsHazelcast 2016-05-19
Test PASSed.