@devOpsHazelcast 2016-06-21
Test PASSed.