@devOpsHazelcast 2016-06-22
Test PASSed.