@devOpsHazelcast 2016-06-27
Test PASSed.